+++ src/controllers/DomainsControllerGetDomainsList.test.ts
@@ -1,15 +1,11 @@
 import supertest from 'supertest';
 import { api } from '../api';
 import { expect } from 'chai';
-import { ApiKey, CnsRegistryEvent, Domain, DomainsResolution } from '../models';
+import { ApiKey, Domain, DomainsResolution } from '../models';
 import { DomainTestHelper } from '../utils/testing/DomainTestHelper';
-import { znsNamehash, eip137Namehash } from '../utils/namehash';
 import { env } from '../env';
-import { getConnection } from 'typeorm';
 import { Blockchain } from '../types/common';
-import { ETHContracts } from '../contracts';
 import { describe } from 'mocha';
-import { DomainAttributes } from './dto/Domains';
 
 describe('DomainsController', () => {
   let testApiKey: ApiKey;
@@ -18,388 +14,6 @@ describe('DomainsController', () => {
     testApiKey = await ApiKey.createApiKey('testing key');
   });
 
-  describe('GET /domain/:domainName', () => {
-    it('should return correct domain resolution for L2 domain on L1', async () => {
-      const { domain } = await DomainTestHelper.createTestDomain({
-        name: 'brad.crypto',
-        node: '0x756e4e998dbffd803c21d23b06cd855cdc7a4b57706c95964a37e24b47c10fc9',
-        ownerAddress: '0x8aaD44321A86b170879d7A244c1e8d360c99DdA8',
-        blockchain: Blockchain.ETH,
-        networkId: 1337,
-        registry: '0xd1e5b0ff1287aa9f9a268759062e4ab08b9dacbe',
-        resolution: {
-          'crypto.ETH.address': '0x8aaD44321A86b170879d7A244c1e8d360c99DdA8',
-        },
-        resolver: '0xd1e5b0ff1287aa9f9a268759062e4ab08b9dacbe',
-      });
-      const resolution = domain.getResolution(Blockchain.MATIC, 1337);
-      resolution.ownerAddress = '0x0000000000000000000000000000000000000000';
-      resolution.resolver = '0xa9a6a3626993d487d2dbda3173cf58ca1a9d9e9f';
-      resolution.registry = '0xa9a6a3626993d487d2dbda3173cf58ca1a9d9e9f';
-      resolution.resolution = {};
-      domain.setResolution(resolution);
-      await domain.save();
-
-      const res = await supertest(api)
-        .get('/domains/brad.crypto')
-        .auth(testApiKey.apiKey, { type: 'bearer' })
-        .send();
-      expect(res.status).eq(200);
-      expect(res.body).containSubset({
-        meta: {
-          domain: 'brad.crypto',
-          owner: '0x8aaD44321A86b170879d7A244c1e8d360c99DdA8',
-          resolver: '0xd1e5b0ff1287aa9f9a268759062e4ab08b9dacbe',
-          registry: '0xd1e5b0ff1287aa9f9a268759062e4ab08b9dacbe',
-          blockchain: 'ETH',
-          networkId: 1337,
-        },
-        records: {
-          'crypto.ETH.address': '0x8aaD44321A86b170879d7A244c1e8d360c99DdA8',
-        },
-      });
-    });
-
-    it('should return correct domain resolution for L2 domain', async () => {
-      const { domain } = await DomainTestHelper.createTestDomain({
-        name: 'brad.crypto',
-        node: '0x756e4e998dbffd803c21d23b06cd855cdc7a4b57706c95964a37e24b47c10fc9',
-        ownerAddress: '0x0000000000000000000000000000000000000000',
-        blockchain: Blockchain.ETH,
-        networkId: 1337,
-        registry: '0xd1e5b0ff1287aa9f9a268759062e4ab08b9dacbe',
-        resolution: {},
-        resolver: '0xd1e5b0ff1287aa9f9a268759062e4ab08b9dacbe',
-      });
-      const resolution = domain.getResolution(Blockchain.MATIC, 1337);
-      resolution.ownerAddress = '0x8aaD44321A86b170879d7A244c1e8d360c99DdA8';
-      resolution.resolver = '0xa9a6a3626993d487d2dbda3173cf58ca1a9d9e9f';
-      resolution.registry = '0xa9a6a3626993d487d2dbda3173cf58ca1a9d9e9f';
-      resolution.resolution = {
-        'crypto.ETH.address': '0x8aaD44321A86b170879d7A244c1e8d360c99DdA8',
-      };
-      domain.setResolution(resolution);
-      await domain.save();
-
-      const res = await supertest(api)
-        .get('/domains/brad.crypto')
-        .auth(testApiKey.apiKey, { type: 'bearer' })
-        .send();
-      expect(res.status).eq(200);
-      expect(res.body).containSubset({
-        meta: {
-          domain: 'brad.crypto',
-          owner: '0x8aaD44321A86b170879d7A244c1e8d360c99DdA8',
-          resolver: '0xa9a6a3626993d487d2dbda3173cf58ca1a9d9e9f',
-          registry: '0xa9a6a3626993d487d2dbda3173cf58ca1a9d9e9f',
-          blockchain: 'MATIC',
-          networkId: 1337,
-        },
-        records: {
-          'crypto.ETH.address': '0x8aaD44321A86b170879d7A244c1e8d360c99DdA8',
-        },
-      });
-    });
-
-    it('should return error for unauthorized query', async () => {
-      const res = await supertest(api).get('/domains/brad.crypto').send();
-      expect(res.status).eq(403);
-      expect(res.body).containSubset({
-        message: 'Please provide a valid API key.',
-      });
-    });
-
-    it('should return non-minted domain', async () => {
-      const res = await supertest(api)
-        .get('/domains/unminted-long-domain.crypto')
-        .auth(testApiKey.apiKey, { type: 'bearer' })
-        .send();
-      expect(res.body).containSubset({
-        meta: {
-          domain: 'unminted-long-domain.crypto',
-          owner: null,
-          resolver: null,
-          registry: null,
-          blockchain: null,
-          networkId: null,
-        },
-        records: {},
-      });
-      expect(res.status).eq(200);
-    });
-
-    it('should return non-minted domain when used a wrong tld', async () => {
-      const res = await supertest(api)
-        .get('/domains/bobby.funnyrabbit')
-        .auth(testApiKey.apiKey, { type: 'bearer' })
-        .send();
-      expect(res.status).eq(200);
-      expect(res.body).containSubset({
-        meta: {
-          domain: 'bobby.funnyrabbit',
-          owner: null,
-          resolver: null,
-          registry: null,
-          blockchain: null,
-          networkId: null,
-        },
-        records: {},
-      });
-    });
-
-    it('should return correct domain resolution for domain in lowercase', async () => {
-      await DomainTestHelper.createTestDomain({
-        name: 'testdomainforcase.crypto',
-        node: '0x08c2e9d2a30aa81623fcc758848d5556696868222fbc80a15ca46ec2fe2cba4f',
-        ownerAddress: '0x8aaD44321A86b170879d7A244c1e8d360c99DdA8',
-        blockchain: Blockchain.ETH,
-        networkId: 1337,
-        registry: '0xd1e5b0ff1287aa9f9a268759062e4ab08b9dacbe',
-        resolution: {
-          'crypto.ETH.address': '0x8aaD44321A86b170879d7A244c1e8d360c99DdA8',
-        },
-        resolver: '0xb66DcE2DA6afAAa98F2013446dBCB0f4B0ab2842',
-      });
-      const res = await supertest(api)
-        .get('/domains/TESTdomainforCase.crypto')
-        .auth(testApiKey.apiKey, { type: 'bearer' })
-        .send();
-      expect(res.status).eq(200);
-      expect(res.body).containSubset({
-        meta: {
-          domain: 'testdomainforcase.crypto',
-          owner: '0x8aaD44321A86b170879d7A244c1e8d360c99DdA8',
-          resolver: '0xb66DcE2DA6afAAa98F2013446dBCB0f4B0ab2842',
-          registry: '0xd1e5b0ff1287aa9f9a268759062e4ab08b9dacbe',
-          blockchain: 'ETH',
-          networkId: 1337,
-        },
-        records: {
-          'crypto.ETH.address': '0x8aaD44321A86b170879d7A244c1e8d360c99DdA8',
-        },
-      });
-    });
-
-    it('should return correct registry for all locations domains', async () => {
-      const { domain: znsDomain } = await DomainTestHelper.createTestDomain({
-        blockchain: Blockchain.ZIL,
-        networkId: env.APPLICATION.ZILLIQA.NETWORK_ID,
-        name: 'test.zil',
-        node: znsNamehash('test.zil'),
-        registry: env.APPLICATION.ZILLIQA.ZNS_REGISTRY_CONTRACT,
-      });
-      const { domain: cnsDomain } = await DomainTestHelper.createTestDomain();
-      const { domain: unsDomain } = await DomainTestHelper.createTestDomain({
-        name: 'test.nft',
-        node: eip137Namehash('test.nft'),
-        registry: ETHContracts.UNSRegistry.address,
-        blockchain: Blockchain.ETH,
-        networkId: env.APPLICATION.ETHEREUM.NETWORK_ID,
-      });
-
-      const znsResult = await supertest(api)
-        .get(`/domains/${znsDomain.name}`)
-        .auth(testApiKey.apiKey, { type: 'bearer' })
-        .send();
-      expect(znsResult.status).eq(200);
-      expect(znsResult.body.meta.registry).eq(
-        env.APPLICATION.ZILLIQA.ZNS_REGISTRY_CONTRACT,
-      );
-
-      const cnsResult = await supertest(api)
-        .get(`/domains/${cnsDomain.name}`)
-        .auth(testApiKey.apiKey, { type: 'bearer' })
-        .send();
-      expect(cnsResult.status).eq(200);
-      expect(cnsResult.body.meta.registry).eq(ETHContracts.CNSRegistry.address);
-
-      const unsResult = await supertest(api)
-        .get(`/domains/${unsDomain.name}`)
-        .auth(testApiKey.apiKey, { type: 'bearer' })
-        .send();
-      expect(unsResult.status).eq(200);
-      expect(unsResult.body.meta.registry).eq(ETHContracts.UNSRegistry.address);
-    });
-
-    it('should return non-minted domain ending on .zil', async () => {
-      const res = await supertest(api)
-        .get('/domains/notreal134522.zil')
-        .auth(testApiKey.apiKey, { type: 'bearer' })
-        .send();
-      expect(res.status).eq(200);
-      expect(res.body).containSubset({
-        meta: {
-          domain: 'notreal134522.zil',
-          owner: null,
-          resolver: null,
-          registry: null,
-          blockchain: null,
-          networkId: null,
-        },
-        records: {},
-      });
-    });
-
-    it('should return minted domain ending on .zil', async () => {
-      await DomainTestHelper.createTestDomain({
-        blockchain: Blockchain.ZIL,
-        networkId: env.APPLICATION.ZILLIQA.NETWORK_ID,
-        name: 'sometestforzil.zil',
-        ownerAddress: '0xcea21f5a6afc11b3a4ef82e986d63b8b050b6910',
-        resolver: '0x34bbdee3404138430c76c2d1b2d4a2d223a896df',
-        registry: '0x9611c53be6d1b32058b2747bdececed7e1216793',
-        node: '0x8052ef7b6b4eee4bc0d7014f0e216db6270bf0055bcd3582368601f2de5e60f0',
-        resolution: {},
-      });
-      const res = await supertest(api)
-        .get('/domains/sometestforzil.zil')
-        .auth(testApiKey.apiKey, { type: 'bearer' })
-        .send();
-      expect(res.status).eq(200);
-      expect(res.body).containSubset({
-        meta: {
-          domain: 'sometestforzil.zil',
-          owner: '0xcea21f5a6afc11b3a4ef82e986d63b8b050b6910',
-          resolver: '0x34bbdee3404138430c76c2d1b2d4a2d223a896df',
-          registry: '0x9611c53be6d1b32058b2747bdececed7e1216793',
-          blockchain: 'ZIL',
-          networkId: 333,
-        },
-        records: {},
-      });
-    });
-
-    it('should return correct domain resolution for minted .crypto domain', async () => {
-      await DomainTestHelper.createTestDomain({
-        name: 'brad.crypto',
-        ownerAddress: '0x8aaD44321A86b170879d7A244c1e8d360c99DdA8',
-        node: '0x756e4e998dbffd803c21d23b06cd855cdc7a4b57706c95964a37e24b47c10fc9',
-        registry: '0xd1e5b0ff1287aa9f9a268759062e4ab08b9dacbe',
-        resolution: {
-          'gundb.username.value':
-            '0x8912623832e174f2eb1f59cc3b587444d619376ad5bf10070e937e0dc22b9ffb2e3ae059e6ebf729f87746b2f71e5d88ec99c1fb3c7c49b8617e2520d474c48e1c',
-          'ipfs.html.value': 'QmdyBw5oTgCtTLQ18PbDvPL8iaLoEPhSyzD91q9XmgmAjb',
-          'ipfs.redirect_domain.value':
-            'https://abbfe6z95qov3d40hf6j30g7auo7afhp.mypinata.cloud/ipfs/Qme54oEzRkgooJbCDr78vzKAWcv6DDEZqRhhDyDtzgrZP6',
-          'crypto.ETH.address': '0x8aaD44321A86b170879d7A244c1e8d360c99DdA8',
-          'gundb.public_key.value':
-            'pqeBHabDQdCHhbdivgNEc74QO-x8CPGXq4PKWgfIzhY.7WJR5cZFuSyh1bFwx0GWzjmrim0T5Y6Bp0SSK0im3nI',
-          'crypto.BTC.address': 'bc1q359khn0phg58xgezyqsuuaha28zkwx047c0c3y',
-        },
-        resolver: '0xb66DcE2DA6afAAa98F2013446dBCB0f4B0ab2842',
-      });
-
-      const res = await supertest(api)
-        .get('/domains/brad.crypto')
-        .auth(testApiKey.apiKey, { type: 'bearer' })
-        .send();
-
-      expect(res.status).eq(200);
-      expect(res.body).containSubset({
-        meta: {
-          domain: 'brad.crypto',
-          owner: '0x8aaD44321A86b170879d7A244c1e8d360c99DdA8',
-          resolver: '0xb66DcE2DA6afAAa98F2013446dBCB0f4B0ab2842',
-          registry: '0xd1e5b0ff1287aa9f9a268759062e4ab08b9dacbe',
-          blockchain: 'ETH',
-          networkId: 1337,
-        },
-        records: {
-          'gundb.username.value':
-            '0x8912623832e174f2eb1f59cc3b587444d619376ad5bf10070e937e0dc22b9ffb2e3ae059e6ebf729f87746b2f71e5d88ec99c1fb3c7c49b8617e2520d474c48e1c',
-          'ipfs.html.value': 'QmdyBw5oTgCtTLQ18PbDvPL8iaLoEPhSyzD91q9XmgmAjb',
-          'ipfs.redirect_domain.value':
-            'https://abbfe6z95qov3d40hf6j30g7auo7afhp.mypinata.cloud/ipfs/Qme54oEzRkgooJbCDr78vzKAWcv6DDEZqRhhDyDtzgrZP6',
-          'crypto.ETH.address': '0x8aaD44321A86b170879d7A244c1e8d360c99DdA8',
-          'gundb.public_key.value':
-            'pqeBHabDQdCHhbdivgNEc74QO-x8CPGXq4PKWgfIzhY.7WJR5cZFuSyh1bFwx0GWzjmrim0T5Y6Bp0SSK0im3nI',
-          'crypto.BTC.address': 'bc1q359khn0phg58xgezyqsuuaha28zkwx047c0c3y',
-        },
-      });
-    });
-  });
-
-  it('should return correct domain resolution for L2 domain', async () => {
-    const { domain } = await DomainTestHelper.createTestDomain({
-      name: 'brad.crypto',
-      node: '0x756e4e998dbffd803c21d23b06cd855cdc7a4b57706c95964a37e24b47c10fc9',
-      ownerAddress: '0x0000000000000000000000000000000000000000',
-      blockchain: Blockchain.ETH,
-      networkId: 1337,
-      registry: '0xd1e5b0ff1287aa9f9a268759062e4ab08b9dacbe',
-      resolution: {},
-      resolver: '0xd1e5b0ff1287aa9f9a268759062e4ab08b9dacbe',
-    });
-    const resolution = domain.getResolution(Blockchain.MATIC, 1337);
-    resolution.ownerAddress = '0x8aaD44321A86b170879d7A244c1e8d360c99DdA8';
-    resolution.resolver = '0xa9a6a3626993d487d2dbda3173cf58ca1a9d9e9f';
-    resolution.registry = '0xa9a6a3626993d487d2dbda3173cf58ca1a9d9e9f';
-    resolution.resolution = {
-      'crypto.ETH.address': '0x8aaD44321A86b170879d7A244c1e8d360c99DdA8',
-    };
-    domain.setResolution(resolution);
-    await domain.save();
-
-    const res = await supertest(api)
-      .get('/domains/brad.crypto')
-      .auth(testApiKey.apiKey, { type: 'bearer' })
-      .send();
-    expect(res.status).eq(200);
-    expect(res.body).containSubset({
-      meta: {
-        domain: 'brad.crypto',
-        owner: '0x8aaD44321A86b170879d7A244c1e8d360c99DdA8',
-        resolver: '0xa9a6a3626993d487d2dbda3173cf58ca1a9d9e9f',
-        registry: '0xa9a6a3626993d487d2dbda3173cf58ca1a9d9e9f',
-        blockchain: 'MATIC',
-        networkId: 1337,
-      },
-      records: {
-        'crypto.ETH.address': '0x8aaD44321A86b170879d7A244c1e8d360c99DdA8',
-      },
-    });
-  });
-
-  it('should return correct domain resolution for L2 domain on L1', async () => {
-    const { domain } = await DomainTestHelper.createTestDomain({
-      name: 'brad.crypto',
-      node: '0x756e4e998dbffd803c21d23b06cd855cdc7a4b57706c95964a37e24b47c10fc9',
-      ownerAddress: '0x8aaD44321A86b170879d7A244c1e8d360c99DdA8',
-      blockchain: Blockchain.ETH,
-      networkId: 1337,
-      registry: '0xd1e5b0ff1287aa9f9a268759062e4ab08b9dacbe',
-      resolution: {
-        'crypto.ETH.address': '0x8aaD44321A86b170879d7A244c1e8d360c99DdA8',
-      },
-      resolver: '0xd1e5b0ff1287aa9f9a268759062e4ab08b9dacbe',
-    });
-    const resolution = await domain.getResolution(Blockchain.MATIC, 1337);
-    resolution.ownerAddress = '0x0000000000000000000000000000000000000000';
-    resolution.resolver = '0xa9a6a3626993d487d2dbda3173cf58ca1a9d9e9f';
-    resolution.registry = '0xa9a6a3626993d487d2dbda3173cf58ca1a9d9e9f';
-    resolution.resolution = {};
-    domain.setResolution(resolution);
-    await domain.save();
-
-    const res = await supertest(api)
-      .get('/domains/brad.crypto')
-      .auth(testApiKey.apiKey, { type: 'bearer' })
-      .send();
-    expect(res.status).eq(200);
-    expect(res.body).containSubset({
-      meta: {
-        domain: 'brad.crypto',
-        owner: '0x8aaD44321A86b170879d7A244c1e8d360c99DdA8',
-        resolver: '0xd1e5b0ff1287aa9f9a268759062e4ab08b9dacbe',
-        registry: '0xd1e5b0ff1287aa9f9a268759062e4ab08b9dacbe',
-        blockchain: 'ETH',
-        networkId: 1337,
-      },
-      records: {
-        'crypto.ETH.address': '0x8aaD44321A86b170879d7A244c1e8d360c99DdA8',
-      },
-    });
-  });
   describe('GET /domains', () => {
     it('should return error for unauthorized query', async () => {
       const res = await supertest(api)
@@ -1134,281 +748,6 @@ describe('DomainsController', () => {
     });
   });
 
-  describe('GET /domains/:domainName/transfers/latest', () => {
-    type eventsTestData = { from: string; to: string }[];
-    async function saveCnsEvents(
-      tokenId: string,
-      l1events: eventsTestData,
-      l2events: eventsTestData,
-    ) {
-      for (let i = 0; i < l1events.length; i++) {
-        const event = l1events[i];
-        await new CnsRegistryEvent({
-          contractAddress: '0xdead1dead1dead1dead1dead1dead1dead1dead1',
-          type: 'Transfer',
-          blockchain: 'ETH',
-          networkId: 1,
-          blockNumber: i,
-          blockHash: `0x${i}`,
-          logIndex: 1,
-          returnValues: { tokenId: tokenId, from: event.from, to: event.to },
-        }).save();
-      }
-      for (let i = 0; i < l2events.length; i++) {
-        const event = l2events[i];
-        await new CnsRegistryEvent({
-          contractAddress: '0xdead2dead2dead2dead2dead2dead2dead2dead2',
-          type: 'Transfer',
-          blockchain: 'MATIC',
-          networkId: 137,
-          blockNumber: i + 1000,
-          blockHash: `0x${i + 1000}`,
-          logIndex: 1,
-          returnValues: { tokenId: tokenId, from: event.from, to: event.to },
-        }).save();
-      }
-    }
-
-    it('should return latest transfers from MATIC and ETH networks', async () => {
-      const { domain: testDomain } = await DomainTestHelper.createTestDomainL2(
-        {
-          name: 'kirill.dao',
-          node: '0x06fd626e68ed0311d37c040c788137dc168124856fdb3b5ec37f54e98dd764ef',
-        },
-        {
-          ownerAddress: '0xea674fdde714fd979de3edf0f56aa9716b898ec8',
-          registry: '0xd1e5b0ff1287aa9f9a268759062e4ab08b9dacbe',
-        },
-        {
-          ownerAddress: '0x58ca45e932a88b2e7d0130712b3aa9fb7c5781e2',
-          registry: '0xd1e5b0ff1287aa9f9a268759062e4ab08b9dacbe',
-        },
-      );
-      await saveCnsEvents(
-        testDomain.node,
-        [
-          {
-            from: '0x0000000000000000000000000000000000000000',
-            to: '0x58ca45e932a88b2e7d0130712b3aa9fb7c5781e2',
-          },
-          {
-            from: '0x58ca45e932a88b2e7d0130712b3aa9fb7c5781e2',
-            to: '0xea674fdde714fd979de3edf0f56aa9716b898ec8',
-          },
-        ],
-        [
-          {
-            from: '0x0000000000000000000000000000000000000000',
-            to: '0x58ca45e932a88b2e7d0130712b3aa9fb7c5781e2',
-          },
-          {
-            from: '0x58ca45e932a88b2e7d0130712b3aa9fb7c5781e2',
-            to: '0xea674fdde714fd979de3edf0f56aa9716b898ec8',
-          },
-          {
-            from: '0xea674fdde714fd979de3edf0f56aa9716b898ec8',
-            to: '0x58ca45e932a88b2e7d0130712b3aa9fb7c5781e2',
-          },
-        ],
-      );
-
-      const res = await supertest(api)
-        .get(`/domains/${testDomain.name}/transfers/latest`)
-        .auth(testApiKey.apiKey, { type: 'bearer' })
-        .send();
-
-      expect(res.status).eq(200);
-      expect(res.body).to.deep.eq({
-        data: [
-          {
-            domain: testDomain.name,
-            from: '0x58ca45e932a88b2e7d0130712b3aa9fb7c5781e2',
-            to: '0xea674fdde714fd979de3edf0f56aa9716b898ec8',
-            blockNumber: 1,
-            networkId: 1,
-            blockchain: 'ETH',
-          },
-          {
-            domain: testDomain.name,
-            from: '0xea674fdde714fd979de3edf0f56aa9716b898ec8',
-            to: '0x58ca45e932a88b2e7d0130712b3aa9fb7c5781e2',
-            blockNumber: 1002,
-            networkId: 137,
-            blockchain: 'MATIC',
-          },
-        ],
-      });
-    });
-
-    it('should return one result if domain has no transfers in another network', async () => {
-      const { domain: testDomain } = await DomainTestHelper.createTestDomainL2(
-        {
-          name: 'kirill.dao',
-          node: '0x06fd626e68ed0311d37c040c788137dc168124856fdb3b5ec37f54e98dd764ef',
-        },
-        {
-          ownerAddress: '0xea674fdde714fd979de3edf0f56aa9716b898ec8',
-          registry: '0xd1e5b0ff1287aa9f9a268759062e4ab08b9dacbe',
-        },
-        {
-          ownerAddress: '0x58ca45e932a88b2e7d0130712b3aa9fb7c5781e2',
-          registry: '0xd1e5b0ff1287aa9f9a268759062e4ab08b9dacbe',
-        },
-      );
-      await saveCnsEvents(
-        testDomain.node,
-        [],
-        [
-          {
-            from: '0x0000000000000000000000000000000000000000',
-            to: '0x58ca45e932a88b2e7d0130712b3aa9fb7c5781e2',
-          },
-          {
-            from: '0x58ca45e932a88b2e7d0130712b3aa9fb7c5781e2',
-            to: '0xea674fdde714fd979de3edf0f56aa9716b898ec8',
-          },
-          {
-            from: '0xea674fdde714fd979de3edf0f56aa9716b898ec8',
-            to: '0x58ca45e932a88b2e7d0130712b3aa9fb7c5781e2',
-          },
-        ],
-      );
-
-      const res = await supertest(api)
-        .get(`/domains/${testDomain.name}/transfers/latest`)
-        .auth(testApiKey.apiKey, { type: 'bearer' })
-        .send();
-
-      expect(res.status).eq(200);
-      expect(res.body).to.deep.eq({
-        data: [
-          {
-            domain: testDomain.name,
-            from: '0xea674fdde714fd979de3edf0f56aa9716b898ec8',
-            to: '0x58ca45e932a88b2e7d0130712b3aa9fb7c5781e2',
-            blockNumber: 1002,
-            networkId: 137,
-            blockchain: 'MATIC',
-          },
-        ],
-      });
-    });
-
-    it('should return error for zil domain', async () => {
-      const res = await supertest(api)
-        .get(`/domains/test.zil/transfers/latest`)
-        .auth(testApiKey.apiKey, { type: 'bearer' })
-        .send();
-
-      expect(res.status).eq(400);
-      expect(res.body.code).to.eq('BadRequestError');
-      expect(res.body.message).to.eq(
-        `Invalid params, check 'errors' property for more info.`,
-      );
-      expect(res.body.errors).to.deep.eq([
-        {
-          children: [],
-          constraints: {
-            'validate domainName with isNotZilDomain': '',
-          },
-          property: 'domainName',
-          target: {
-            domainName: 'test.zil',
-          },
-          value: 'test.zil',
-        },
-      ]);
-    });
-  });
-
-  describe('Errors handling', () => {
-    it('should format the 500 error', async () => {
-      const connection = getConnection();
-      connection.close();
-      const res = await supertest(api)
-        .get('/domains/brad.crypto')
-        .auth(testApiKey.apiKey, { type: 'bearer' })
-        .send();
-      expect(res.status).eq(500);
-      expect(res.body.code).to.exist;
-      expect(res.body.message).to.exist;
-      expect(res.body.errors).to.exist;
-      expect(res.body.stack).to.not.exist;
-      await connection.connect(); // restore the connection to the db;
-    });
-    it('should return appropriate error for missing an owner param', async () => {
-      const res = await supertest(api)
-        .get('/domains/')
-        .auth(testApiKey.apiKey, { type: 'bearer' })
-        .send();
-      expect(res.status).eq(400);
-      expect(res.body).containSubset({
-        code: 'BadRequestError',
-        message: "Invalid queries, check 'errors' property for more info.",
-        errors: [
-          {
-            property: 'owners',
-            constraints: {
-              arrayNotEmpty: 'owners should not be empty',
-              isArray: 'owners must be an array',
-              isNotEmpty: 'each value in owners should not be empty',
-              isString: 'each value in owners must be a string',
-            },
-          },
-        ],
-      });
-    });
-    it('should return appropriate error for incorrect input', async () => {
-      const res = await supertest(api)
-        .get(
-          '/domains?owners[]=0x58ca45e932a88b2e7d0130712b3aa9fb7c5781e2&perPage=0',
-        )
-        .auth(testApiKey.apiKey, { type: 'bearer' })
-        .send();
-
-      expect(res.status).eq(400);
-      expect(res.body.code).to.exist;
-      expect(res.body.message).to.exist;
-      expect(res.body).to.containSubset({
-        code: 'BadRequestError',
-        message: "Invalid queries, check 'errors' property for more info.",
-        errors: [
-          {
-            property: 'perPage',
-            constraints: {
-              min: 'perPage must not be less than 1',
-            },
-          },
-        ],
-      });
-    });
-
-    it('should return error for incorrect tlds', async () => {
-      const res = await supertest(api)
-        .get(
-          '/domains?owners[]=0x58ca45e932a88b2e7d0130712b3aa9fb7c5781e2&tlds=crypto&tlds=test',
-        )
-        .auth(testApiKey.apiKey, { type: 'bearer' })
-        .send();
-
-      expect(res.status).eq(400);
-      expect(res.body.code).to.exist;
-      expect(res.body.message).to.exist;
-      expect(res.body).to.containSubset({
-        code: 'BadRequestError',
-        message: "Invalid queries, check 'errors' property for more info.",
-        errors: [
-          {
-            property: 'tlds',
-            constraints: {
-              'validate tlds with validTlds': 'Invalid TLD list provided',
-            },
-          },
-        ],
-      });
-    });
-  });
-
   describe('GET /domains sorting and filtration', () => {
     let testDomains: {
       domain: Domain;
@@ -1636,6 +975,58 @@ describe('DomainsController', () => {
       });
     });
 
+    it('should sort by created_at ascending', async () => {
+      const { domains, expectedData } = getSortedTestDomains(
+        (a, b) =>
+          (a.domain.createdAt?.getDate() || 0) -
+          (b.domain.createdAt?.getDate() || 0),
+      );
+
+      const res = await supertest(api)
+        .get(
+          `/domains?owners[]=0x58ca45e932a88b2e7d0130712b3aa9fb7c5781e2&sortBy=created_at&sortDirection=ASC`,
+        )
+        .auth(testApiKey.apiKey, { type: 'bearer' })
+        .send();
+
+      expect(res.status).eq(200);
+      expect(res.body.data).to.exist;
+      expect(res.body.data).to.deep.equal(expectedData);
+      expect(res.body.meta).to.deep.equal({
+        hasMore: false,
+        nextStartingAfter: domains[domains.length - 1].createdAt?.toString(),
+        perPage: 100,
+        sortBy: 'created_at',
+        sortDirection: 'ASC',
+      });
+    });
+
+    it('should sort by created_at descending', async () => {
+      const { domains, expectedData } = getSortedTestDomains(
+        (a, b) =>
+          (b.domain.createdAt?.getDate() || 0) -
+          (a.domain.createdAt?.getDate() || 0),
+      );
+
+      const res = await supertest(api)
+        .get(
+          `/domains?owners[]=0x58ca45e932a88b2e7d0130712b3aa9fb7c5781e2&sortBy=created_at&sortDirection=DESC`,
+        )
+        .auth(testApiKey.apiKey, { type: 'bearer' })
+        .send();
+
+      expect(res.status).eq(200);
+      expect(res.body.data).to.exist;
+      expect(res.body.data).to.deep.equal(expectedData);
+      expect(res.body.meta).to.deep.equal({
+        hasMore: false,
+        nextStartingAfter: domains[domains.length - 1].createdAt?.toString(),
+        perPage: 100,
+        sortBy: 'created_at',
+        sortDirection: 'DESC',
+      });
+    });
+
     it('should sort with starting after', async () => {
       const { domains, expectedData } = getSortedTestDomains(
         (a, b) => -a.domain.name.localeCompare(b.domain.name),
@@ -1694,6 +1085,58 @@ describe('DomainsController', () => {
           {
             constraints: {
               isIn: 'sortDirection must be one of the following values: ASC, DESC',
+            },
+          },
+        ],
+      });
+    });
+  });
+
+  describe('Errors handling', () => {
+    it('should return appropriate error for incorrect input', async () => {
+      const res = await supertest(api)
+        .get(
+          '/domains?owners[]=0x58ca45e932a88b2e7d0130712b3aa9fb7c5781e2&perPage=0',
+        )
+        .auth(testApiKey.apiKey, { type: 'bearer' })
+        .send();
+
+      expect(res.status).eq(400);
+      expect(res.body.code).to.exist;
+      expect(res.body.message).to.exist;
+      expect(res.body).to.containSubset({
+        code: 'BadRequestError',
+        message: "Invalid queries, check 'errors' property for more info.",
+        errors: [
+          {
+            property: 'perPage',
+            constraints: {
+              min: 'perPage must not be less than 1',
+            },
+          },
+        ],
+      });
+    });
+
+    it('should return error for incorrect tlds', async () => {
+      const res = await supertest(api)
+        .get(
+          '/domains?owners[]=0x58ca45e932a88b2e7d0130712b3aa9fb7c5781e2&tlds=crypto&tlds=test',
+        )
+        .auth(testApiKey.apiKey, { type: 'bearer' })
+        .send();
+
+      expect(res.status).eq(400);
+      expect(res.body.code).to.exist;
+      expect(res.body.message).to.exist;
+      expect(res.body).to.containSubset({
+        code: 'BadRequestError',
+        message: "Invalid queries, check 'errors' property for more info.",
+        errors: [
+          {
+            property: 'tlds',
+            constraints: {
+              'validate tlds with validTlds': 'Invalid TLD list provided',
             },
           },
         ],
